+++ src/http/hooks/logs.hooks.js
@@ -11,3 +11,15 @@ export async function logRequestsHandler({ event, resolve }) {
 
 	return await resolve(event);
 }
+
+/** @type {import('@sveltejs/kit').Handle} */
+export async function logRequestDetailsHandler({ event, resolve }) {
+	/**
+	 * Logging incoming request details
+	 * TODO users and sessions will be added later
+	 */
+
+	logger.info(JSON.stringify(event));
+
+	return await resolve(event);
+}