+++ src/http/status/index.js
@@ -6,3 +6,8 @@ export const http_500 = {
 	status: 500,
 	text: 'Internal Server Error'
 };
+
+export const http_404 = {
+	status: 404,
+	text: 'Not Found'
+};